backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy
aiosqlite==0.21.0
pydantic[email]==2.11.7
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.20
python-dotenv==1.1.1
email-validator